requirements.txt
# Note: Make .env file and add gemini api key (GEMINI_API_KEY=AIza...) before running the project

# Install: 

streamlit
python-dotenv
sqlparse
pandas
google-generativeai
sqlalchemy
